refactor: rename package to com.postcare.app
