+++ app/src/main/AndroidManifest.xml
@@ -10,12 +10,12 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.MyApplication">
+        android:theme="@style/Theme.PostCare">
 
         <activity
             android:name=".HomePage"
             android:exported="true"
-            android:theme="@style/Theme.MyApplication.NoActionBar">
+            android:theme="@style/Theme.PostCare.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER"/>
