+++ app/src/main/java/com/postcare/app/PatientAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.postcare.app
 
 import android.view.LayoutInflater
 import android.view.View
