+++ app/src/main/java/com/postcare/app/Question.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.postcare.app
 
 data class Question(
     val text: String,
