+++ app/src/main/java/com/postcare/app/ui/home/HomeFragment.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication.ui.home
+package com.postcare.app.ui.home
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,7 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.ViewModelProvider
-import com.example.myapplication.databinding.FragmentHomeBinding
+import com.postcare.app.databinding.FragmentHomeBinding
 
 class HomeFragment : Fragment() {
