+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,25 +7,25 @@
 
     <fragment
         android:id="@+id/nav_home"
-        android:name="com.example.myapplication.ui.home.HomeFragment"
+        android:name="com.postcare.app.ui.home.HomeFragment"
         android:label="@string/menu_home"
         tools:layout="@layout/fragment_home" />
 
     <fragment
         android:id="@+id/nav_gallery"
-        android:name="com.example.myapplication.ui.gallery.GalleryFragment"
+        android:name="com.postcare.app.ui.gallery.GalleryFragment"
         android:label="@string/menu_gallery"
         tools:layout="@layout/fragment_gallery" />
 
     <fragment
         android:id="@+id/nav_slideshow"
-        android:name="com.example.myapplication.ui.slideshow.SlideshowFragment"
+        android:name="com.postcare.app.ui.slideshow.SlideshowFragment"
         android:label="@string/menu_slideshow"
         tools:layout="@layout/fragment_slideshow" />
 
     <fragment
         android:id="@+id/preparationFragment"
-        android:name="com.example.myapplication.ui.PreparationFragment"
+        android:name="com.postcare.app.ui.PreparationFragment"
         android:label="Préparation">
         <action
             android:id="@+id/action_preparation_to_questionnaire"
@@ -34,7 +34,7 @@
 
     <fragment
         android:id="@+id/questionnaireFragment"
-        android:name="com.example.myapplication.ui.QuestionnaireFragment"
+        android:name="com.postcare.app.ui.QuestionnaireFragment"
         android:label="Questionnaire" />
 
 </navigation>
