+++ app/src/test/java/com/postcare/app/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.postcare.app
 
 import org.junit.Test
